move SubjectIndex under annif.vocab where it more properly belongs
