+++ annif/vocab.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import csv
 import os.path
 from typing import TYPE_CHECKING
 
@@ -15,10 +16,11 @@ if TYPE_CHECKING:
     from rdflib.graph import Graph
 
     from annif.corpus.skos import SubjectFileSKOS
-    from annif.corpus.subject import SubjectCorpus, SubjectIndex
+    from annif.corpus.subject import Subject, SubjectCorpus
 
 
 logger = annif.logger
+logger.addFilter(annif.util.DuplicateFilter())
 
 
 class AnnifVocabulary(DatadirMixin):
@@ -39,16 +41,16 @@ class AnnifVocabulary(DatadirMixin):
         self._skos_vocab = None
 
     def _create_subject_index(self, subject_corpus: SubjectCorpus) -> SubjectIndex:
-        subjects = annif.corpus.SubjectIndex()
+        subjects = SubjectIndex()
         subjects.load_subjects(subject_corpus)
         annif.util.atomic_save(subjects, self.datadir, self.INDEX_FILENAME_CSV)
         return subjects
 
     def _update_subject_index(self, subject_corpus: SubjectCorpus) -> SubjectIndex:
         old_subjects = self.subjects
-        new_subjects = annif.corpus.SubjectIndex()
+        new_subjects = SubjectIndex()
         new_subjects.load_subjects(subject_corpus)
-        updated_subjects = annif.corpus.SubjectIndex()
+        updated_subjects = SubjectIndex()
 
         for old_subject in old_subjects:
             if new_subjects.contains_uri(old_subject.uri):
@@ -70,7 +72,7 @@ class AnnifVocabulary(DatadirMixin):
             path = os.path.join(self.datadir, self.INDEX_FILENAME_CSV)
             if os.path.exists(path):
                 logger.debug("loading subjects from %s", path)
-                self._subjects = annif.corpus.SubjectIndex.load(path)
+                self._subjects = SubjectIndex.load(path)
             else:
                 raise NotInitializedException("subject file {} not found".format(path))
         return self._subjects
@@ -158,3 +160,108 @@ class AnnifVocabulary(DatadirMixin):
             "size": size,
             "loaded": loaded,
         }
+
+
+class SubjectIndex:
+    """An index that remembers the associations between integers subject IDs
+    and their URIs and labels."""
+
+    def __init__(self) -> None:
+        self._subjects = []
+        self._uri_idx = {}
+        self._label_idx = {}
+        self._languages = None
+
+    def load_subjects(self, corpus: SubjectCorpus) -> None:
+        """Initialize the subject index from a subject corpus"""
+
+        self._languages = corpus.languages
+        for subject in corpus.subjects:
+            self.append(subject)
+
+    def __len__(self) -> int:
+        return len(self._subjects)
+
+    @property
+    def languages(self) -> list[str] | None:
+        return self._languages
+
+    def __getitem__(self, subject_id: int) -> Subject:
+        return self._subjects[subject_id]
+
+    def append(self, subject: Subject) -> None:
+        if self._languages is None and subject.labels is not None:
+            self._languages = list(subject.labels.keys())
+
+        subject_id = len(self._subjects)
+        self._uri_idx[subject.uri] = subject_id
+        if subject.labels:
+            for lang, label in subject.labels.items():
+                self._label_idx[(label, lang)] = subject_id
+        self._subjects.append(subject)
+
+    def contains_uri(self, uri: str) -> bool:
+        return uri in self._uri_idx
+
+    def by_uri(self, uri: str, warnings: bool = True) -> int | None:
+        """return the subject ID of a subject by its URI, or None if not found.
+        If warnings=True, log a warning message if the URI cannot be found."""
+        try:
+            return self._uri_idx[uri]
+        except KeyError:
+            if warnings:
+                logger.warning("Unknown subject URI <%s>", uri)
+            return None
+
+    def by_label(self, label: str | None, language: str) -> int | None:
+        """return the subject ID of a subject by its label in a given
+        language"""
+        try:
+            return self._label_idx[(label, language)]
+        except KeyError:
+            logger.warning('Unknown subject label "%s"@%s', label, language)
+            return None
+
+    def deprecated_ids(self) -> list[int]:
+        """return indices of deprecated subjects"""
+
+        return [
+            subject_id
+            for subject_id, subject in enumerate(self._subjects)
+            if subject.labels is None
+        ]
+
+    @property
+    def active(self) -> list[tuple[int, Subject]]:
+        """return a list of (subject_id, subject) tuples of all subjects that
+        are not deprecated"""
+
+        return [
+            (subj_id, subject)
+            for subj_id, subject in enumerate(self._subjects)
+            if subject.labels is not None
+        ]
+
+    def save(self, path: str) -> None:
+        """Save this subject index into a file with the given path name."""
+
+        fieldnames = ["uri", "notation"] + [f"label_{lang}" for lang in self._languages]
+
+        with open(path, "w", encoding="utf-8", newline="") as csvfile:
+            writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
+            writer.writeheader()
+            for subject in self:
+                row = {"uri": subject.uri, "notation": subject.notation or ""}
+                if subject.labels:
+                    for lang, label in subject.labels.items():
+                        row[f"label_{lang}"] = label
+                writer.writerow(row)
+
+    @classmethod
+    def load(cls, path: str) -> SubjectIndex:
+        """Load a subject index from a CSV file and return it."""
+
+        corpus = annif.corpus.SubjectFileCSV(path)
+        subject_index = cls()
+        subject_index.load_subjects(corpus)
+        return subject_index
