+++ tests/test_corpus.py
@@ -226,26 +226,6 @@ def test_docdir_key_as_doccorpus(tmpdir, subject_index):
     )
 
 
-def test_subject_by_uri(subject_index):
-    subj_id = subject_index.by_uri("http://www.yso.fi/onto/yso/p7141")
-    assert subject_index[subj_id].labels["fi"] == "sinetit"
-
-
-def test_subject_by_uri_missing(subject_index):
-    subj_id = subject_index.by_uri("http://nonexistent")
-    assert subj_id is None
-
-
-def test_subject_by_label(subject_index):
-    subj_id = subject_index.by_label("sinetit", "fi")
-    assert subject_index[subj_id].uri == "http://www.yso.fi/onto/yso/p7141"
-
-
-def test_subject_by_label_missing(subject_index):
-    subj_id = subject_index.by_label("nonexistent", "fi")
-    assert subj_id is None
-
-
 def test_docfile_plain(tmpdir, subject_index):
     docfile = tmpdir.join("documents.tsv")
     docfile.write(
